% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_scores.R
\name{get_data_scores}
\alias{get_data_scores}
\title{get_data_scores}
\usage{
get_data_scores(
  OTU_Table,
  NMDS = c(c("NMDS1", "NMDS2"), c("NMDS1", "NMDS3"), c("NMDS2", "NMDS3")),
  logtransform = T,
  relative = T,
  type
)
}
\arguments{
\item{OTU_Table}{an OTU table including sample metadata in columns 1-5}

\item{NMDS}{which NMDS axes to display. Must give two, e.g. c("NMDS1", "NMDS2")}

\item{logtransform}{logical, should abundances be log transformed?}

\item{relative}{logical, should relative abundances be used?}

\item{type}{one of "Microhabitat" or "TreeSpecies". Returns hull data for either}
}
\value{
a dataframe containing the convex hull coordinates for the given microhabitats
}
\description{
Performs NMDS ordination and saves convex hulls for all habitats in hull.data
}
\examples{

data(OTU_Table)
get_data_scores(OTU_Table = OTU_Table, NMDS = c("NMDS1", "NMDS2"), 
relative = T, logtransform = T, type = "Microhabitat")

}
