% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_scores2.R
\name{get_data_scores2}
\alias{get_data_scores2}
\title{get_data_scores2}
\usage{
get_data_scores2(
  OTU_Table,
  NMDS = c(c("NMDS1", "NMDS2"), c("NMDS1", "NMDS3"), c("NMDS2", "NMDS3")),
  logtransform = T,
  relative = T,
  habitat = c("Microhabitat", "TreeSpecies"),
  which = c("all", c("Habitat1", "Habitat2", "..."))
)
}
\arguments{
\item{OTU_Table}{an OTU table including sample metadata in columns 1-5}

\item{NMDS}{which NMDS axes to display. Must specify two, e.g. c("NMDS1", "NMDS2")}

\item{logtransform}{logical, should abundances be log transformed?}

\item{relative}{logical, should relative abundances be used?}

\item{habitat}{which sample type to draw hull data from. Must be a column name in the sample metadata}

\item{which}{which samples to use. Must be "all" or vector of specified habitats (see details)}
}
\value{
a dataframe containing the convex hull coordinates for the given habitats
}
\description{
Performs NMDS ordination and saves convex hulls for specified habitats in \code{hull.data}
}
\examples{

data(OTU_Table)
get_data_scores(OTU_Table = OTU_Table, NMDS = c("NMDS1", "NMDS2"), 
relative = T, logtransform = T, habitat = "Microhabitat", which = "all")

}
