% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hull_data.R
\name{get_hull_data}
\alias{get_hull_data}
\title{get_hull_data}
\usage{
get_hull_data(
  OTU_Table,
  SampleMetadata,
  NMDS = c(c("NMDS1", "NMDS2"), c("NMDS1", "NMDS3"), c("NMDS2", "NMDS3")),
  logtransform = T,
  relative = T,
  habitat = c("Microhabitat", "TreeSpecies"),
  which = c("all", c("Habitat1", "Habitat2", "..."))
)
}
\arguments{
\item{OTU_Table}{an OTU table with or without sample metadata}

\item{SampleMetadata}{the Metadata for the samples. Can be separate table or
specified which columns in the OTU table should be used (start and end, e.g. c(1, 5)).
Must be in beginning of OTU table}

\item{NMDS}{which NMDS axes to display.
Must specify two, e.g. c("NMDS1", "NMDS2")}

\item{logtransform}{logical, should abundances be log transformed?}

\item{relative}{logical, should relative abundances be used?}

\item{habitat}{which sample type to draw hull data from.
Must be a column name in the sample metadata}

\item{which}{which samples to use.
Must be "all" or vector of specified habitats (see details)}
}
\value{
a dataframe containing the convex hull coordinates for the given habitats
}
\description{
Performs NMDS ordination and saves convex hulls for specified habitats
}
\examples{

data(OTU_Table)

hull.data = get_hull_data(
  OTU_Table = OTU_Table, 
  SampleMetadata = c(1, 5),
  NMDS = c("NMDS1", "NMDS2"), 
  relative = TRUE, 
  logtransform = TRUE, 
  habitat = "Microhabitat", 
  which = "all")

}
