% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NMDS_Ordination.R
\name{get_NMDS_ordination}
\alias{get_NMDS_ordination}
\title{get_NMDS_ordination}
\usage{
get_NMDS_ordination(OTU_Table, SampleMetadata, logtransform = T, relative = T)
}
\arguments{
\item{OTU_Table}{an OTU table with or without sample metadata}

\item{SampleMetadata}{the Metadata for the samples. Can be separate table or
specified which columns in the OTU table should be used (start and end, e.g. c(1, 5)).
Must be in beginning of OTU table}

\item{logtransform}{logical, should abundances be log transformed?}

\item{relative}{logical, should relative abundances be used?}
}
\value{
a dataframe containing NMDS ordination
}
\description{
Performs NMDS ordination and saves it in a dataframe
}
\examples{

data(OTU_Table)

NMDS.data = get_NMDS_ordination(
  OTU_Table = OTU_Table, 
  SampleMetadata = c(1, 5),
  relative = TRUE, 
  logtransform = TRUE)

}
