% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OTU_Table}
\alias{OTU_Table}
\title{OTU_Table}
\format{
A data frame with several variables:

\describe{
\item{\code{SampleID}}{Unique sample identificatior}
\item{\code{TreeID}}{Unique tree identificatior}
\item{\code{TreeSpecies}}{Species of sampled tree}
\item{\code{Microhabitat}}{Sampled Microhabitat}
\item{\code{Stratum}}{Sampled Stratum}
\item{\code{X...}}{OTU Number and taxonomy}
}
}
\usage{
OTU_Table
}
\description{
Abundance table of OTUs in Microhabitats
}
\keyword{datasets}
