% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NMDS.R
\name{plot_NMDS}
\alias{plot_NMDS}
\title{plot_NMDS}
\usage{
plot_NMDS(hull_data, hull_data2, x, y, group1, group2)
}
\arguments{
\item{hull_data}{the result of the \code{get_hull_data} function}

\item{hull_data2}{additional hull data if you want to plot several types}

\item{x}{first NMDS axis to display, e.g. "NMDS1"}

\item{y}{second NMDS axis to display, e.g. "NMDS2"}

\item{group1}{which grouping variable to use, e.g. "Microhabitat"}

\item{group2}{second grouping variable, e.g. "TreeSpecies"}
}
\value{
a ggplot object with the convex hull as polygons
}
\description{
Plots NMDS ordination based on convex hull data generated from \code{get_hull_data}
}
\examples{
data(OTU_Table)

hull.data = get_hull_data(
  OTU_Table = OTU_Table, 
  SampleMetadata = c(1, 5),
  NMDS = c("NMDS1", "NMDS2"), 
  relative = TRUE, 
  logtransform = TRUE, 
  habitat = "Microhabitat", 
  which = "all")

hull.data.TreeSpecies = get_hull_data(
  OTU_Table = OTU_Table, 
  SampleMetadata = c(1, 5),
  NMDS = c("NMDS1", "NMDS2"), 
  relative = TRUE, 
  logtransform = TRUE, 
  habitat = "TreeSpecies", 
  which = "all")
  
g = plot_NMDS(hull_data = hull.data, hull_data2 = hull.data.TreeSpecies, 
  x = "NMDS1", y = "NMDS2", 
  group1 = "Microhabitat", group2 = "TreeSpecies")

}
