% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NMDS.R
\name{plot_NMDS}
\alias{plot_NMDS}
\title{plot_NMDS}
\usage{
plot_NMDS(hull_data, hull_data2 = NULL, NMDS_data, x, y, habitat,
  habitat2 = NULL)
}
\arguments{
\item{hull_data}{the result of the \code{\link[otutableNMDS]{get_hull_data}} function}

\item{hull_data2}{optional additional hull data if you want to plot several types}

\item{NMDS_data}{the result of the \code{\link[otutableNMDS]{get_NMDS_ordination}} function}

\item{x}{first NMDS axis to display, e.g. "NMDS1"}

\item{y}{second NMDS axis to display, e.g. "NMDS2"}

\item{habitat}{which grouping variable to use, e.g. "Microhabitat".
Must be the same habitat used in the \code{\link[otutableNMDS]{get_hull_data}} function}

\item{habitat2}{optional second grouping variable, e.g. "TreeSpecies"}
}
\value{
a ggplot object with the convex hull as polygons
}
\description{
Plots NMDS ordination based on convex hull data generated from \code{get_hull_data}
}
\examples{
data(OTU_Table)

NMDS.data = get_NMDS_ordination(
  OTU_Table = OTU_Table, 
  SampleMetadata = c(1, 5),
  relative = TRUE, 
  logtransform = TRUE, 
  k = 3, 
  trymax = 100)

hull.data = get_hull_data(
  NMDS_data = NMDS.data, 
  SampleMetadata = OTU_Table[,1:5],
  NMDS = c("NMDS1", "NMDS2"), 
  habitat = "Microhabitat", 
  which = "all")

hull.data.TreeSpecies = get_hull_data(
  NMDS_data = NMDS.data, 
  SampleMetadata = OTU_Table[,1:5],
  NMDS = c("NMDS1", "NMDS2"), 
  habitat = "TreeSpecies", 
  which = "all")
  
g = plot_NMDS(
  hull_data = hull.data, 
  hull_data2 = hull.data.TreeSpecies, 
  NMDS_data = NMDS.data,
  x = "NMDS1", 
  y = "NMDS2", 
  habitat = "Microhabitat", 
  habitat2 = "TreeSpecies")

}
